requirements.txt
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
python-dotenv==1.1.0
pytest==8.3.5
pytest-django==4.11.1
gunicorn==23.0.0
django-redis==5.4.0
redis==6.2.0
